// src/beatmap/interfaces/beatmap-general.interface.ts
export interface IBeatmapGeneral {
  length: number;
  bpmMin: number;
  bpmMax: number;
  bpmMode: number;
  circleSize: number;
  approachRate: number;
  overallDifficulty: number;
  drainRate: number;
  hittable: number;
  slidable: number;
  spinnable: number;
  holdable: number;
  maxCombo: number;
  totalHits: number;
}